Add files via upload
coffee shop website project is a sleek and user-friendly frontend designed to mimic the experience of a food delivery website. Built using HTML, CSS, and JavaScript, it allows users to browse various coffee options, customize their orders, and view the menu with ease. The site aims to provide an interactive and visually appealing interface that enhances the online coffee shopping experience, similar to popular delivery platforms.